requirements.txt
opencv-python
numpy
torch
tensorflow
keras
safetensors
transformers
python-dotenv
requests